requirements.txt
requests>=2.31.0
pytest>=7.4.0
python-dotenv>=1.0.0
pylint>=3.0.0
coverage>=7.0.0